+++ tests/test_core_methods.py
@@ -1,4 +1,5 @@
 import pytest
+from datetime import datetime
 
 from pykwalify.core import Core
 from pykwalify.errors import NotSequenceError
@@ -16,3 +17,55 @@ def test_validate_sequence():
     with pytest.raises(NotSequenceError):
         c = Core(source_data={}, schema_data={})
         c._validate_sequence(123, Rule(sequence=['']), '', [])
+
+
+def test_validate_timestamp():
+    c = Core(source_data={}, schema_data={})
+
+    errors = []
+    c._validate_scalar_timestamp("", errors, '')
+    assert len(errors) == 1
+
+    errors = []
+    c._validate_scalar_timestamp("1234567", errors, '')
+    assert len(errors) == 0
+
+    errors = []
+    c._validate_scalar_timestamp("2016-01-01", errors, '')
+    assert len(errors) == 0
+
+    errors = []
+    c._validate_scalar_timestamp("2016-01-01 15:01", errors, '')
+    assert len(errors) == 0
+
+    errors = []
+    c._validate_scalar_timestamp(123, errors, '')
+    assert len(errors) == 0
+
+    errors = []
+    c._validate_scalar_timestamp(1.5, errors, '')
+    assert len(errors) == 0
+
+    errors = []
+    c._validate_scalar_timestamp(0, errors, '')
+    assert len(errors) == 1
+
+    errors = []
+    c._validate_scalar_timestamp(-1, errors, '')
+    assert len(errors) == 1
+
+    errors = []
+    c._validate_scalar_timestamp(3147483647, errors, '')
+    assert len(errors) == 1
+
+    errors = []
+    c._validate_scalar_timestamp([], errors, '')
+    assert len(errors) == 1
+
+    errors = []
+    c._validate_scalar_timestamp(datetime.now(), errors, '')
+    assert len(errors) == 0
+
+    errors = []
+    c._validate_scalar_timestamp(datetime.today(), errors, '')
+    assert len(errors) == 0